Refactor changelog from slide-out panel tab to full-page main view
- Extract ChangelogTab from DigestPanel into standalone ChangelogView component
- Add ChangelogView as a regular view in App (same pattern as home/library/collections)
- Rename bottom tab from 'Digest' to 'Changelog', treat it as a normal view tab
- Add Changelog button to desktop nav bar
- Remove DigestPanel component and all digestOpen state/logic from App
- Simplify PlaybackBar toggle and desktop padding (no more digest panel offset)
